Parallelize copy and download, download in chunks
We now perform downloads for the next chunk of data during the time that
Copy is loading the previous chunk of data into the database. This will
entirely hide the cost of downloading and improves overall performance
significantly.

The download task downloads chunks of the file by making
ranged GET requests. The chunk size chosen after experimentation is 100MiB.

Authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>
Authored-by: Karen Huddleston <khuddleston@pivotal.io>