Proper tool dependency and go1.12- compatibility
From the go1.13 release notes:
The GO111MODULE environment variable continues to default to auto, but
the auto setting now activates the module-aware mode of the go command
whenever the current working directory contains, or is below a directory
containing, a go.mod file — even if the current directory is within
GOPATH/src.

While we transition to our new dependency manager it will be a good idea
to maintain a bit of backward compatibility with older version of
golang. This change does not affect developers compiling with go1.13+,
but fixes compilation for developers compiling with go1.12-.

Added tooling dependency

Authored-by: Kevin Yeap <kyeap@pivotal.io>